requirement.txt
streamlit
pypdf2
langchain
python-dotenv
faiss-cpu
openai
huggingface-hub==0.14.1
InstructorEmbedding==1.0.1
sentence-transformers==2.2.2
InstructorEmbedding